+++ example/test/widget_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:stories_example/main.dart';
+import 'package:storify_example/main.dart';
 
 void main() {
   testWidgets('app loads', (tester) async {